ros_gz_bridge in ROS2-Humble does not use declare_paramters.
Therefore bridges cannot be configured from launch files by passing bridge config to the parameters field of the Node
action. We have to pass a file in the parameters field. And this file must have the bridge config.
It's not clear how to use composition as it is in ROS2-Jazzy, so for the time being composition is removed.